[REFACTOR] 필요없는 파일 제거
