--- src/main/java/store/ckin/auth/config/RestTemplateConfig.java
@@ -1,25 +1,0 @@
-package store.ckin.auth.config;
-
-import java.time.Duration;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-/**
- * RestTemplate 설정 클래스 입니다.
- *
- * @author : jinwoolee
- * @version : 2024. 02. 19.
- */
-@Configuration
-public class RestTemplateConfig {
-
-    @Bean
-    RestTemplate restTemplate(RestTemplateBuilder builder) {
-        return builder
-                .setReadTimeout(Duration.ofSeconds(5L))
-                .setConnectTimeout(Duration.ofSeconds(5L))
-                .build();
-    }
-}
