--- src/main/java/store/ckin/auth/member/exception/ServerErrorException.java
@@ -1,10 +1,0 @@
-package store.ckin.auth.member.exception;
-
-/**
- * 내부 서버 에러가 감지됐을 때 호출 되는 Exception 클래스 입니다.
- *
- * @author : jinwoolee
- * @version : 2024. 03. 04.
- */
-public class ServerErrorException extends RuntimeException{
-}
