+++ src/main/java/store/ckin/auth/token/service/TokenService.java
@@ -1,7 +1,7 @@
 package store.ckin.auth.token.service;
 
-import store.ckin.auth.token.service.domain.TokenRequestDto;
-import store.ckin.auth.token.service.domain.TokenResponseDto;
+import store.ckin.auth.token.domain.TokenRequestDto;
+import store.ckin.auth.token.domain.TokenResponseDto;
 
 /**
  * Token 관리 로직을 처리하는 Service interface 입니다.
