--- src/main/resources/application-dev.properties
@@ -1,1 +1,0 @@
-spring.redis.database=12
