+++ src/main/resources/application.properties
@@ -1,12 +1,6 @@
 # Auth Server port = 7010
 server.port=7010
 
-# Gateway Port
-port.gateway-uri=http://localhost:9010
-
-# API Port
-port.api-uri = http://localhost:7030
-
 # Redis
 spring.cache.type=redis
 spring.redis.host=133.186.241.167
